Add toString / inspect to Exit
