+++ .changeset/green-tables-run.md
@@ -1,0 +1,5 @@
+---
+"@effect/io": patch
+---
+
+Add toString / inspect to Exit
