+++ src/internal/runtime.ts
@@ -11,7 +11,7 @@ import type * as RuntimeFlags from "@effect/io/Fiber/Runtime/Flags"
 import type * as FiberRef from "@effect/io/FiberRef"
 import * as FiberRefs from "@effect/io/FiberRefs"
 import { NoSuchElementException } from "@effect/io/internal/cause"
-import * as CausePretty from "@effect/io/internal/cause-pretty"
+import * as CausePretty from "@effect/io/internal/cause"
 import * as core from "@effect/io/internal/core"
 import * as FiberRuntime from "@effect/io/internal/fiberRuntime"
 import * as fiberScope from "@effect/io/internal/fiberScope"
